Moved non-cargo related code to the new `geiger` crate.
