+++ Cargo.toml
@@ -1,25 +1,6 @@
-[package]
-name = "cargo-geiger"
-description = "Detects usage of unsafe Rust in your project and its dependencies."
-version = "0.4.2"
-edition = "2018"
-authors = ["anderejd <rajder@gmail.com>"]
-repository = "https://github.com/anderejd/cargo-geiger"
-readme = "README.md"
-keywords = ["unsafe", "cargo", "plugin"]
-categories = ["command-line-utilities", "development-tools", "development-tools::cargo-plugins", "parsing"]
-license = "Apache-2.0/MIT"
+[workspace]
 
-[badges]
-maintenance = { status = "experimental" }
-
-[dependencies]
-syn = { version = "0.15.26", features = ["parsing", "printing", "clone-impls", "full", "extra-traits", "visit"] }
-clap = "2.32.0"
-cargo = "0.33.0"
-env_logger = "0.6.0"
-failure = "0.1.5"
-petgraph = "0.4.13"
-structopt = "0.2.14"
-walkdir = "2.2.7"
-colored = "1.7.0"
+members = [
+    "cargo-geiger",
+    "geiger",
+]
