+++ geiger/.gitignore
@@ -4,7 +4,7 @@
 
 # Remove Cargo.lock from gitignore if creating an executable, leave it for libraries
 # More information here http://doc.crates.io/guide.html#cargotoml-vs-cargolock
-# Cargo.lock
+Cargo.lock
 
 # These are backup files generated by rustfmt
 **/*.rs.bk
