+++ geiger/Cargo.toml
@@ -1,0 +1,18 @@
+[package]
+name = "geiger"
+description = "Some library parts of cargo-geiger, decoupled from cargo."
+version = "0.1.0"
+edition = "2018"
+authors = ["anderejd <rajder@gmail.com>"]
+repository = "https://github.com/anderejd/cargo-geiger"
+readme = "README.md"
+keywords = ["unsafe"]
+categories = ["development-tools", "parsing"]
+license = "Apache-2.0/MIT"
+
+[badges]
+maintenance = { status = "experimental" }
+
+[dependencies]
+syn = { version = "0.15.26", features = ["parsing", "printing", "clone-impls", "full", "extra-traits", "visit"] }
+walkdir = "2.2.7"
