+++ geiger/README.md
@@ -1,0 +1,21 @@
+geiger ☢
+========
+
+This crate provides some library parts used by `cargo-geiger` that are decoupled
+from `cargo`.
+
+For more details please see the README.md in `cargo-geiger`.
+
+TODO
+----
+
+Flesh out this readme.
+
+Changelog
+---------
+
+### Unreleased (0.1.0)
+ - Parts of `cargo-geiger` was moved to this crate.
+
+
+
